Fix test selection of optimized/naive impl
Should now work with all combinations of feature flags,
properly indicating when the optimized and naive impls are being used.
Before the implementations were being attributed the wrong names,
assuming the default was still the opimized impl.
Now the default is the naive impl, and the tests have been changed
to reflect that.

Also fixes usage of feature="prefer-optimized-impl"